// src/profile_utils/schema.ts
export interface BaseObject {
  Id: string;
  Name?: string;
  DeveloperName?: string;
  SobjectType?: string;
  FullName?: string;
  NamespacePrefix?: string;
  ManageableState?: string;
  attributes?: {
    type: string;
    url: string;
  };
}

export interface RecordType extends BaseObject {
  Description?: string;
  IsActive?: boolean;
  IsPersonType?: boolean;
  BusinessProcessId?: string;
}

export interface EntityDefinition extends BaseObject {
  DurableId?: string;
  QualifiedApiName?: string;
}

export interface Layout extends BaseObject {
  ObjectName?: string;
  EntityDefinitionId?: string;
}